# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Contracts API",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/",
            "description": "Base server"
        }
    ],
    "paths": {
...
        "/api/contracts": {
            "get": {
                "tags": [
                    "Contracts"
                ],
                "summary": "კონტრაქტების სიის ნახვა",
                "operationId": "d8997f490da9062b7b77f68eaa735f07",
                "parameters": [
                    {
                        "name": "party_name",
                        "in": "query",
                        "description": "ძიება მხოლოდ სახელით (party_name)",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "description": "სტატუსი (მაგ: აქტივი, დასამტკიცებელი და სხვ.)",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "sign_date_from",
                        "in": "query",
                        "description": "ხელმოწერის თარიღი - საწყისი",
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "sign_date_to",
                        "in": "query",
                        "description": "ხელმოწერის თარიღი - საბოლოო",
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "expiry_date_from",
                        "in": "query",
                        "description": "ვადის თარიღი - საწყისი",
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "expiry_date_to",
                        "in": "query",
                        "description": "ვადის თარიღი - საბოლოო",
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "sort_by",
                        "in": "query",
                        "description": "სორტირების ველი",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "id",
                                "contract_number",
                                "sign_date",
                                "expiry_date",
                                "amount",
                                "created_at",
                                "updated_at"
                            ]
                        }
                    },
                    {
                        "name": "sort_dir",
                        "in": "query",
                        "description": "სორტირების მიმართულება",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "asc",
                                "desc"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Contracts"
                ],
                "summary": "კონტრაქტის შექმნა",
                "operationId": "674f25b6428b177ae2a1fe1123b4dcf6",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "party_name",
                                    "party_identifier",
                                    "contract_type_id",
                                    "subject",
                                    "sign_date",
                                    "branch_id",
                                    "currency",
                                    "status",
                                    "payment_type"
                                ],
                                "properties": {
                                    "party_name": {
                                        "description": "მეორე მხარის სახელი (legacy)",
                                        "type": "string",
                                        "example": "შპს ალფა"
                                    },
                                    "party_identifier": {
                                        "description": "იდენტიფიკატორი 9-11 ციფრი",
                                        "type": "string",
                                        "example": "123456789"
                                    },
                                    "contractor_id": {
                                        "description": "კონტრაქტორის ID (ნებაყოფლობითი)",
                                        "type": "integer",
                                        "example": 1,
                                        "nullable": true
                                    },
                                    "contract_number": {
                                        "description": "ხელშეკრულების ნომერი (უნიკალური)",
                                        "type": "string",
                                        "example": "CNT-2025-0003",
                                        "nullable": true
                                    },
                                    "contract_type_id": {
                                        "description": "ხელშეკრულების ტიპი",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "subject": {
                                        "description": "საგანი/შინაარსი",
                                        "type": "string",
                                        "example": "სერვისების მიწოდება და მხარდაჭერა"
                                    },
                                    "sign_date": {
                                        "description": "ხელმოწერის თარიღი",
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-09-10"
                                    },
                                    "expiry_date": {
                                        "description": "ვადის თარიღი (>= sign_date)",
                                        "type": "string",
                                        "format": "date",
                                        "example": "2026-09-10",
                                        "nullable": true
                                    },
                                    "branch_id": {
                                        "description": "ფილიალი",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "department_id": {
                                        "description": "დეპარტამენტი",
                                        "type": "integer",
                                        "example": 1,
                                        "nullable": true
                                    },
                                    "currency": {
                                        "description": "ვალუტა ISO 4217",
                                        "type": "string",
                                        "example": "GEL"
                                    },
                                    "amount": {
                                        "description": "თანხა",
                                        "type": "number",
                                        "format": "float",
                                        "example": 12000,
                                        "nullable": true
                                    },
                                    "status": {
                                        "description": "სტატუსი (default: დასამტკიცებელი)",
                                        "type": "string",
                                        "example": "დასამტკიცებელი"
                                    },
                                    "responsible_manager_id": {
                                        "description": "პასუხისმგებელი მენეჯერი (users.id)",
                                        "type": "integer",
                                        "example": null,
                                        "nullable": true
                                    },
                                    "payment_type": {
                                        "description": "გადახდის ტიპი",
                                        "type": "string",
                                        "example": "ერთჯერადი"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/contracts/{id}/attachments": {
            "get": {
                "tags": [
                    "Contracts"
                ],
                "summary": "კონტრაქტზე მიმაგრებული ფაილების სია",
                "operationId": "97e2438635edc463044a324ff7d61cf5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Contracts"
                ],
                "summary": "ფაილების ატვირთვა კონტრაქტზე (მხარდაჭერა მრავალში)",
                "operationId": "8fe40b851ac0b5931a80dc57770eccd0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "files[]": {
                                        "description": "ერთზე მეტი ფაილი (pdf/doc/docx/xlsx/csv, ≤50MB თითოეული)",
                                        "type": "array",
                                        "items": {
                                            "type": "string",
                                            "format": "binary"
                                        }
                                    },
                                    "file": {
                                        "description": "ერთეული ფაილი (ალტერნატიული ველი)",
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Uploaded"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/contracts/export": {
            "get": {
                "tags": [
                    "Contracts"
                ],
                "summary": "კონტრაქტების ექსპორტი CSV/XLSX/PDF",
                "operationId": "eb237ab3346231149b869ec53daab646",
                "parameters": [
                    {
                        "name": "format",
                        "in": "query",
                        "description": "ექსპორტის ფორმატი (ფილტრებიც მოქმედებს)",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "csv",
                                "xlsx",
                                "pdf"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "File stream"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}